Fix(auth_handler,manager_handler): При удалении менеджера, функционал менеджера у пользователя исчезает до момента, пока его не добавят в менеджеры

+++ src/handlers/admin_hand/manager_handler.py
@@ -19,24 +19,42 @@ from graph_service.graph_service import (
     statistic_citys_graph,
 )
 
+from database.database_manager import get_user_role_from_manager
+
 router = Router()
 
 @router.message(F.text == "График новых пользователей", ManagerState.MANAGER_STATE)
 @router.message(F.text == "График новых пользователей", AdminState.ADMIN_STATE)
 async def new_user_graphic_handler(message: Message):
+    username = message.from_user.username
+    role = await get_user_role_from_manager(username)
+
+    if role not in ["manager", "admin"]:
+        await message.answer("❌ У вас нет доступа к этой функции.")
+        return
+
     graph_path = await new_users_graph()
-
-    await message.answer_photo(photo=FSInputFile(graph_path), caption="График новых пользователей")
-
+    await message.answer_photo(
+        photo=FSInputFile(graph_path),
+        caption="График новых пользователей"
+    )
     os.remove(graph_path)
 
 
 @router.message(F.text == "График популярности городов", ManagerState.MANAGER_STATE)
 @router.message(F.text == "График популярности городов", AdminState.ADMIN_STATE)
 async def workload_graphic_handler(message: Message):
+    username = message.from_user.username
+    role = await get_user_role_from_manager(username)
+
+    if role not in ["manager", "admin"]:
+        await message.answer("❌ У вас нет доступа к этой функции.")
+        return
+
     graph_path = await statistic_citys_graph()
-
-    await message.answer_photo(photo=FSInputFile(graph_path), caption="График популярности городов")
-
+    await message.answer_photo(
+        photo=FSInputFile(graph_path),
+        caption="График популярности городов"
+    )
     os.remove(graph_path)
